Added code to handle queries with too many tokens and also lowered the weight of important words
If query has more than 32 tokens, serve query based on only the 32 longest tokens. This change was made because serving query scales with the size of the query, so if the query reaches a certain, it will eventually exceed 300 ms. A limit has to be placed how how many tokens we should consider when parsing the query. Also, changed the weight of important words because it feels more fitting that an important words weighs twice as much as a normal word rather than three times.